Fix boundaries issues

+++ src/jeuDeLaVie/Coordinate.kt
@@ -1,6 +1,6 @@
 package jeuDeLaVie
 
-class Coordinate (val x: Int, val y: Int, val z: Int) {
+class Coordinate(val x: Double, val y: Double, val z: Double) {
     override fun toString(): String {
         return "(x=$x, y=$y, z=$z)"
     }
